Fix clippy::type_complexity, refactor type alias
Use a type alias to refactor the submit_funcs type definition, as a
vector of boxed SubmitFuncs. Note this type alias isn't used in the
parameterized add_submit_func() function because it accepts a trait, not
a type - no `dyn` keyword in this context, but it is needed in the type
definition.